feat: define is_slot_conflicted function
Checking slots conflicted or not to prevent time overlap with existing
bookings for a staff member. Basically in previous case services were
independently generated slots for each of them while there was one staff
was doing all the service. In this logic it prevents overlaping with
different services. Better time managing system.